rename painter file name
